moving grid-specific methods to torch_utils
