--- grid/services/__init__.py
@@ -1,6 +1,0 @@
-from . import *
-from . import base
-from . import fit_worker
-from . import listen_for_openmined_nodes
-from . import broadcast_known_workers
-from . import torch
