+++ grid/workers/compute.py
@@ -3,8 +3,8 @@ from ..lib import strings, utils, output_pipe
 from .. import channels
 from ..services.fit_worker import FitWorkerService
 from ..services.listen_for_openmined_nodes import ListenForOpenMinedNodesService
-from ..services.torch.torch_service import TorchService
-from ..services.torch.hook_worker_service import HookWorkerService
+from ..services.torch_services.torch_service import TorchService
+from ..services.torch_services.hook_worker_service import HookWorkerService
 import json
 import threading
